requirements.txt
azure-core==1.8.2
azure-storage-blob==12.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
idna==2.10
isodate==0.6.0
msrest==0.6.19
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
